+++ donations/admin.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8
-from __future__ import unicode_literals, absolute_import
-
 from django.contrib import admin
 
 from donations.models import Donation, Frequency, DonationProvider